requirements.txt
antlr4-python3-runtime==4.9.3
asttokens==2.4.1
certifi==2023.7.22
charset-normalizer==3.3.2
cmake==3.27.7
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.1
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.17.3
hydra-core==1.3.2
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
lit==17.0.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
packaging==23.2
pandas==2.1.2
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
